Expose labels and annotations on the template and parent resource
Reworked the opinions to accept the resource instead of the
PodTemplateSpec contained within the resource. This makes the resource's
annotations and labels available to be managed.

A Resource interface limits the exposure of other fields on the
SpringBootApplication resource, as well as maintains decoupling between
the opinions and the SpringBootApplication to allow for other types of
resources and sets of opinions.